refactor: index.js 함수 구조 변경
raceCarGame 함수 내에서 뷰 초기화, 이벤트핸들러 처리
initView함수 view 디렉토리로 이동
inputEventHandler 함수면 수정 및 input 디렉토리로 이동
