+++ src/view/initView.js
@@ -1,0 +1,8 @@
+import View from './View.js';
+
+export default function initView() {
+  const view = new View();
+
+  view.hideNextHtml();
+  return view;
+}
